materials_discovery/requirements.txt
ase
jupyterlab
nglview
numba
numpy
paddlepaddle-gpu==1.10
paddlepaddle-gpu-scatter
scipy>=1.3
sympy>=1.5
tqdm
